Reformulate in terms of preserving dependencies
Reformulate the paper's conception of semantic dependency relative to
a particular compiler.  In this new version, the text talks instead
about compiled code preserving syntactic dependencies.

In the original paper, perhaps the most questionable aspect was the
unspoken issue of whether the definition of semantic dependency did an
adequate job of capturing the informal notion.  In the new version,
this issue translates into an explicit Dependency Preservation thesis,
stating that semantic dependencies are always preserved.

In addition to the virtue of bringing this core issue to the
forefront, the new approach may have a psychological advantage.
People naturally tend to think that the key characteristic of semantic
dependencies -- the thing that makes them truly semantic -- is that
they _must_ be preserved, there is no way to avoid or destroy them.

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>