ootavacuous: Volatile accesses *might* synchronize
The current text is correct when it says that non-benign environments can
destroy synchronization, for example, by storing to memory.  However, if
the environment shared normal memory with the I/O firmware, and if this
firmware happened to avoid doing such a store during the time interval
of interest to any of the variables of interest, then the normal atomic
synchronization would apply (though it would of course not be guaranteed
by the standard).

So add a few more occurrences of the word "might".

Signed-off-by: Paul E. McKenney <paulmck@kernel.org>